added check guest logic into CRUDjs

+++ client/src/crud/CRUD.js
@@ -1,7 +1,13 @@
 import axios from 'axios'
+import {store} from './../index.js'
 
-export function getRandomPic(user){
+//console.log(store.getState().user.user);
+
+
+export function getRandomPic(){
+
   return new Promise((resolve,reject)=>{
+    let user = store.getState().user.user.username==="Guest" ? store.getState().user.user.userip : store.getState().user.user.username
     axios.get('/api/crud/'+user)
     .then((response)=>{
       resolve(response.data)
@@ -12,7 +18,12 @@ export function getRandomPic(user){
   })
 }
 
-export function getProfilePics(user){
+export function getProfilePics(){
+  if(store.getState().user.user.username==="Guest"){
+   console.log("Guests not allowed here");
+   return null;
+  }
+  const user = store.getState().user.user.username;
   return new Promise((resolve,reject)=>{
     axios.get('/api/crud/profile/'+user)
       .then((response)=>{
@@ -25,6 +36,13 @@ export function getProfilePics(user){
 }
 
 export function createPic(picInfo){
+  if(store.getState().user.user.username==="Guest"){
+    console.log("Guests not allowed here")
+    return null;
+  }
+
+  picInfo.owner = store.getState().user.user.username;
+
   return new Promise((resolve,reject)=>{
     axios.post('/api/crud',picInfo)
       .then((response)=>{
@@ -49,6 +67,11 @@ export function updatePic(picID,updateQuery){
 }
 
 export function deletePic(picID){
+
+  if(store.getState().user.user.username==="Guest"){
+    console.log("Guests not allowed here")
+    return null;
+  }
   return new Promise((resolve,reject)=>{
     axios.delete('/api/crud/'+picID)
     .then((response)=>{
